Hibernate banking application : fix bug

+++ hibernate/src/main/java/com/poc/hibernate/dto/ClientCompteCountDto.java
@@ -1,6 +1,6 @@
 package com.poc.hibernate.dto;
 
 public record ClientCompteCountDto(
-        Long id, String nom, double nombreComptes
+        Long id, String nom, Long nombreComptes
 ) {
 }
